fix: use correct toast ID for transformation completion
The transformation completion was using the wrong toast ID variable,
causing the "Running transformation..." loading spinner to persist
indefinitely. It was using the original recording's toastId instead
of the transformToastId created specifically for the transformation
operation.

This fix ensures the transformation loading toast is properly replaced
when the transformation completes.


Former-commit-id: cc5c86428b00839d477869f89bf32b133fae5585 [formerly 741f1b3ea801ae2083afe8b843458245ed1df2a3 [formerly 741f1b3ea801ae2083afe8b843458245ed1df2a3 [formerly 39dd15694d89d96579e7c54a351cdf3c0155074b]]]
Former-commit-id: f7dd38a42a12e6b947addaa79a703ec4595f2a41
Former-commit-id: 3cc1d610383de9e68afbdc6cb8e03106eaf4b70d
Former-commit-id: 2061295ef8da1aed94ed96f340f63afad2611851